svc/slanger/requirements.txt
Flask==2.3.3
Flask-Monitoring==1.1.2
prometheus-client==0.20.0
Jinja2==3.1.2
MarkupSafe==2.1.3
PyYAML==6.0.1
Werkzeug==2.3.7
alpaca-py==0.17.0
annotated-types==0.6.0
blinker==1.6.2
boto3==1.28.44
botocore==1.31.44
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
idna==3.4
itsdangerous==2.1.2
jmespath==1.0.1
msgpack==1.0.5
numpy==1.25.2
pandas==2.1.0
prometheus-flask-exporter==0.23.1
pydantic==2.6.3
pydantic_core==2.16.3
pyngrok==7.1.6
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sseclient-py==1.8.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
websockets==11.0.3
coloredlogs==15.0.1
influxdb-client==1.46.0 
pifx==1.0.1 